reorganize utility functions and tests
